remove win10 build warnings in detect stage

+++ modules/barcode/src/detector/bardetect.hpp
@@ -55,7 +55,7 @@ protected:
 
     static inline bool isValidCoord(const Point &coord, const Size &limit);
 
-    static inline double computeOrientation(float y, float x);
+    static inline float computeOrientation(float y, float x);
 
     void regionGrowing(int window_size);
